% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{add_attr_from_tdr}
\alias{add_attr_from_tdr}
\title{add CODEC attributes to a data.frame based on a tabular-data-resource list}
\usage{
add_attr_from_tdr(.x, tdr)
}
\arguments{
\item{.x}{a data.frame or tibble}

\item{tdr}{a tabular-data-resource list (usually created with \code{read_tdr()})}
}
\value{
.x with added tabular-data-resource attributes
}
\description{
add CODEC attributes to a data.frame based on a tabular-data-resource list
}
