% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{add_attrs}
\alias{add_attrs}
\alias{add_col_attrs}
\title{add attributes}
\usage{
add_attrs(.x, ...)

add_col_attrs(.x, var, ...)
}
\arguments{
\item{.x}{an object to add attributes to}

\item{...}{additional arguments with name and value corresponding to metadata name and value}

\item{var}{name of variable in .x to add attributes to}
}
\value{
.x with the updated attributes
}
\description{
\itemize{
\item Use \code{add_attrs} to add attributes to an R object
\item Use \code{add_col_attrs} to add attributes to a column inside of a data frame
}
}
\examples{
add_attrs(mtcars, name = "Motor Trend Cars", year = "1974")
add_col_attrs(mtcars, mpg, name = "MPG", description = "Miles Per Gallon")
}
