% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_census_tract_id}
\alias{check_census_tract_id}
\title{Check for census tract id column}
\usage{
check_census_tract_id(.x)
}
\arguments{
\item{.x}{a codec tabular-data-resource}
}
\value{
.x, invisibly
}
\description{
Errors will be raised if the CODEC tabular data resource does not meet the following requirements:
\itemize{
\item data MUST contain a census tract identifier column titled
\code{census_tract_id_2000}, \code{census_tract_id_2010}, or \code{census_tract_id_2020}
\item census tract identifier column MUST contain all census tract identifiers
in Hamilton County, OH for the appropriate vintage
}
}
