% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_codec_tdr_csv}
\alias{check_codec_tdr_csv}
\title{Check a tabular-data-resource against CODEC specifications}
\usage{
check_codec_tdr_csv(.x)
}
\arguments{
\item{.x}{path to folder containing the tdr}
}
\value{
tibble with added tabular-data-resource attributes
(equivalent to read_tdr_csv with \code{codec = TRUE})
}
\description{
This function will check that the file structure is correct,
that the metadata can be read using CODEC specifications,
that the CSV data can be read in with accompanying metadata,
and that the data contains a census tract identifier column.
}
