% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_tdr_name}
\alias{check_tdr_name}
\title{check CODEC tdr name}
\usage{
check_tdr_name(name)
}
\arguments{
\item{name}{the name field from tabular-data-resource.yaml}
}
\value{
NULL, invisibly
}
\description{
Name must be an identifier string composed of lower
case alphanumeric characters, _, -, and .
}
\keyword{internal}
