% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_census_geo}
\alias{cincy_census_geo}
\title{Cincy census tracts and block groups}
\usage{
cincy_census_geo(
  geography = c("tract", "bg"),
  vintage = as.character(2024:2013),
  packaged = TRUE
)
}
\arguments{
\item{geography}{which type of cincy census geography to return}

\item{vintage}{a character vector of a year corresponding to the
vintage of TIGER/Line data}

\item{packaged}{logical; use the data included with the package instead
of (down)loading from the source data?}
}
\value{
a simple features object with a geographic identifier column
(\code{geoid}) and a geometry column (\code{s2_geography})
}
\description{
Read tract and block group ("bg") geographies from the online Census
TIGER/Line files into R.
}
\details{
Compressed shapefiles are downloaded from TIGER into an R user data
directory and will be cached for use across other R sessions.
}
\examples{
cincy_census_geo("tract", "2020")
cincy_census_geo("bg", "2020")
}
