% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_neighborhood_geo}
\alias{cincy_neighborhood_geo}
\title{Cincy neighborhood geographies}
\usage{
cincy_neighborhood_geo(
  geography = c("statistical_neighborhood_approximations", "community_council"),
  packaged = TRUE
)
}
\arguments{
\item{geography}{which type of cincy neighborhood geography to return}

\item{packaged}{logical; use the data included with the package instead
of (down)loading from the source data?}
}
\value{
a simple features object with a geographic identifier column
(\code{geoid}) and a geometry column (\code{s2_geography})
}
\description{
CAGIS data (see \code{install_cagis_data()}) provides community council
boundaries, but these boundaries can overlap and do not align with
census geographies or ZIP codes.
By default, the statistical neighborhood approximations are instead
returned, which are calculated by aggregating census tracts into 50
matching neighborhoods.
}
\examples{
cincy_neighborhood_geo()
}
