% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_zcta_geo}
\alias{cincy_zcta_geo}
\title{Cincy ZIP Code Tabulation Areas}
\usage{
cincy_zcta_geo(vintage = as.character(2024:2013), packaged = TRUE)
}
\arguments{
\item{vintage}{a character vector of a year corresponding to the
vintage of TIGER/Line data}

\item{packaged}{logical; use the data included with the package instead
of (down)loading from the source data?}
}
\value{
a simple features object with a geographic identifier column
(\code{geoid}) and a geometry column (\code{s2_geography})
}
\description{
Read ZIP Code Tabulation Areas (ZCTAs) geographies from the online Census
TIGER/Line files into R.
}
\examples{
cincy_zcta_geo("2020")
}
