% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pins.R
\name{codec_board}
\alias{codec_board}
\title{CoDEC pin board hosted on GitHub}
\usage{
codec_board(cache, use_cache_on_failure, headers)
}
\arguments{
\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{use_cache_on_failure}{If the pin fails to download, is it ok to
use the last cached version? Defaults to \code{is_interactive()} so you'll
be robust to poor internet connectivity when exploring interactively,
but you'll get clear errors when the code is deployed.}

\item{headers}{Named character vector for additional HTTP headers (such as for
authentication). See \code{\link[pins:connect_auth_headers]{connect_auth_headers()}} for Posit Connect support.}
}
\value{
a pins_board object
}
\description{
A shortcut to create a pins board using the url of the CoDEC pins
board manifest file on GitHub.
Use this function with the pins package to get CoDEC data tables
from the online CoDEC data catalog.
}
\examples{
library(pins)
pin_list(codec_board())
pin_read(codec_board(), "property_code_enforcements")
pin_meta(codec_board(), "property_code_enforcements")
}
