% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg.R
\name{codec_dpkg_add_geog}
\alias{codec_dpkg_add_geog}
\title{Add a simple features geometry column to a CoDEC data package}
\usage{
codec_dpkg_add_geog(x)
}
\arguments{
\item{x}{a codec data package}
}
\value{
a codec data package with an additional simple features geometry column
}
\description{
The required census tract identifier column name is used to merge in tract geographies
from the \href{https://geomarker.io/cincy}{cincy} package.
}
\examples{
get_codec_dpkg("drivetime-v0.2.2") |>
  codec_dpkg_add_geog()
}
