% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg.R
\name{codec_dpkg_as_sf}
\alias{codec_dpkg_as_sf}
\title{Convert a CoDEC dpkg into an sf object}
\usage{
codec_dpkg_as_sf(x)
}
\arguments{
\item{x}{a CoDEC dpkg}
}
\value{
an \code{sf} object that is a codec dpkg with an added \code{geometry} column
}
\description{
The required census tract identifier column name is used to merge in tract geographies
from the \href{https://geomarker.io/cincy}{cincy} package.
}
\examples{
get_codec_dpkg("drivetime-v0.2.2") |>
  codec_dpkg_as_sf()
}
