% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{codec_names}
\alias{codec_names}
\title{valid names for CODEC descriptor, schema, and fields}
\usage{
codec_names()
}
\value{
a list of character vectors, one for each of "descriptor", "schema", and "fields"
}
\description{
valid names for CODEC descriptor, schema, and fields
}
\examples{
codec_names()
}
