% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{codec_tdr}
\alias{codec_tdr}
\title{codec_tdr}
\usage{
codec_tdr()
}
\value{
a list of named character vectors, one for each of "descriptor", "schema", and "fields"
}
\description{
Defines the list of names (and descriptions) of valid CODEC descriptor, schema, and fields
}
\examples{
names(codec_tdr())
}
