% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODECtools-package.R
\docType{package}
\name{CODECtools-package}
\alias{CODECtools}
\alias{CODECtools-package}
\title{CODECtools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geomarker-io/CODECtools}
  \item Report bugs at \url{https://github.com/geomarker-io/CODECtools/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID})

}
\keyword{internal}
