% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_gh.R
\name{dpkg_gh_release}
\alias{dpkg_gh_release}
\title{Use a dpkg to create a github release}
\usage{
dpkg_gh_release(x, draft = TRUE)
}
\arguments{
\item{x}{path to datapackage.yaml}

\item{draft}{logical; mark release as draft?}
}
\value{
the URL to the release (invisibly)
}
\description{
The release will be tagged at the current commit and named
according to the \code{name} and \code{version} of the dpkg.
Any \code{description} metadata (auto populated at time of dpkg creation using a README.md file)
will also be used as the GitHub release description.
The \code{GITHUB_PAT} environment variable must be set and the working directory
must be inside of a git repository with a GitHub remote.
Trying to create more than one release from the current commit will result in an error.
}
