% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg.R
\name{dpkg_read}
\alias{dpkg_read}
\title{Read a data package from a folder}
\usage{
dpkg_read(dpkg, readRDS = TRUE)
}
\arguments{
\item{dpkg}{path or url to folder containing a datapackage.yaml file}

\item{readRDS}{logical; read and return R object based on \code{path} descriptor?}
}
\value{
if readRDS is TRUE, then an R object saved read from RDS file;
otherwise, the path to the data resource
}
\description{
Read a \href{https://datapackage.org/standard/data-package/}{data package} with
one \href{https://datapackage.org/standard/data-resource/}{data resource} that
is a RDS object created with \code{dpkg_write()}.
}
\examples{
dpkg_write(mtcars, "mtcars", version = "0.1.0", dir = tempdir())
dpkg_read(fs::path(tempdir(), "mtcars"))
}
