% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg.R
\name{dpkg_s3_put}
\alias{dpkg_s3_put}
\title{Put a dpkg file into the CoDEC S3 bucket under \verb{/data}}
\usage{
dpkg_s3_put(x)
}
\arguments{
\item{x}{path or url to datapackage.yaml file}
}
\description{
Ensure {paws} can connect via the usual methods
}
