% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg.R
\name{get_codec_dpkg}
\alias{get_codec_dpkg}
\title{Read a dpkg from the public CoDEC repository into R}
\usage{
get_codec_dpkg(codec_dpkg, overwrite = FALSE)
}
\arguments{
\item{codec_dpkg}{name of CoDEC dpkg}

\item{overwrite}{logical; re-download the remote file even though
a local file with the same name exists?}
}
\value{
a CoDEC data package (see \code{dpkg::as_dpkg()})
}
\description{
Public data packages are downloaded from \verb{gh://geomarker-io/codec/} using
\code{dpkg::stow()} to cache a local copy in the user's data directory.
}
