% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_descriptors}
\alias{get_descriptors}
\title{get descriptors from attributes of data frame}
\usage{
get_descriptors(.x)
}
\arguments{
\item{.x}{data frame or tibble}
}
\value{
a tibble with \code{name} and \code{value} columns for each attribute
}
\description{
\code{get_descriptors()} looks for attributes with a value of
a single character string and returns them in a tibble.
Other attributes will not be returned.
To get complete data resource metadata in a list,
use \code{make_data_resource_from_attr()}
}
