% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{glimpse_attr}
\alias{glimpse_attr}
\title{glimpse attributes of a data.frame}
\usage{
glimpse_attr(.x, codec = TRUE)
}
\arguments{
\item{.x}{data frame or tibble}

\item{codec}{logical; include only CODEC properties? (see \code{?codec_tdr} for details)}
}
\value{
a tibble with \code{name} and \code{value} columns of attributes
}
\description{
glimpse attributes of a data.frame
}
\details{
values are collapsed using \verb{,}
}
