% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{glimpse_schema}
\alias{glimpse_schema}
\title{glimpse CoDEC schema}
\usage{
glimpse_schema(.x)
}
\arguments{
\item{.x}{data frame or tibble}
}
\value{
a tibble with attributes for each column in \code{.x}
}
\description{
glimpse CoDEC schema
}
\details{
constraints are collapsed using \verb{,}
}
