% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{make_attr_from_tdr}
\alias{make_attr_from_tdr}
\title{add attributes in a metadata list to a data.frame}
\usage{
make_attr_from_tdr(.x, metadata)
}
\arguments{
\item{.x}{a data.frame or tibble}

\item{metadata}{a list containing frictionless metadata (usually created with \code{read_tdr()})}
}
\value{
.x with added frictionless metadata attributes
}
\description{
add attributes in a metadata list to a data.frame
}
