% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-resource.R
\name{make_data_resource_from_attr}
\alias{make_data_resource_from_attr}
\title{make metadata list from attributes of data.frame}
\usage{
make_data_resource_from_attr(.x)
}
\arguments{
\item{.x}{a data.frame or tibble}
}
\value{
a list of metadata generated using the attributes of .x
}
\description{
This also includes schema based on the attributes of columns.
This function is largely for use in \code{write_metadata()} and
should normally not need to be used directly.
}
