% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_metadata_from_attr}
\alias{make_metadata_from_attr}
\title{make metadata list based on attributes of data.frame (including schema based on attributes of columns)}
\usage{
make_metadata_from_attr(.x)
}
\description{
make metadata list based on attributes of data.frame (including schema based on attributes of columns)
}
