% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{make_tdr_from_attr}
\alias{make_tdr_from_attr}
\title{make a tabular-data-resource list from the attributes of a data.frame}
\usage{
make_tdr_from_attr(.x, codec = TRUE)
}
\arguments{
\item{.x}{a data.frame or tibble}

\item{codec}{logical; use only CODEC descriptors?}
}
\value{
a list of tabular-data-resource metadata
}
\description{
make a tabular-data-resource list from the attributes of a data.frame
}
