% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{read_tdr}
\alias{read_tdr}
\title{read metadata in from a tabular-data-resource.yaml file}
\usage{
read_tdr(file = "tabular-data-resource.yaml")
}
\arguments{
\item{file}{filename (or connection) of yaml file to read metadata from}
}
\value{
a list of frictionless metadata
}
\description{
read metadata in from a tabular-data-resource.yaml file
}
