% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-resource.R
\name{read_tdr_csv}
\alias{read_tdr_csv}
\title{read a CSV file and frictionless metadata into R}
\usage{
read_tdr_csv(file = "tabular-data-resource.yaml")
}
\description{
read a CSV file and frictionless metadata into R
}
