% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{release_codec_tdr}
\alias{release_codec_tdr}
\title{experimental!: release a CODEC tdr}
\usage{
release_codec_tdr(.x, version, s3 = FALSE, gh = FALSE)
}
\arguments{
\item{.x}{a CODEC tdr}

\item{version}{the version number to be used for the release}

\item{s3}{logical; create and upload CODEC tdr to s3://codec-data/ ?}

\item{gh}{logical; create and upload CODEC tdr to a new GitHub release?}
}
\description{
If releasing to AWS S3, the CSV data file will be uploaded first and its
VersionId from AWS S3 metadata will be added to the properties of the uploaded
tabular-data-resource.yaml file
}
