% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-resource.R
\name{save_tdr}
\alias{save_tdr}
\title{extract data resource metadata from a data frame and save it to a file}
\usage{
save_tdr(.x, file = "tabular-data-resource.yaml")
}
\arguments{
\item{.x}{a data.frame or tibble}

\item{file}{name of yaml file to write metadata to}
}
\value{
.x (invisibly)
}
\description{
extract data resource metadata from a data frame and save it to a file
}
\examples{
\dontrun{
mtcars |>
  add_attrs(name = "Motor Trend Cars", year = "1974") |>
  add_type_attrs() |>
  add_col_attrs(mpg, name = "MPG", description = "Miles Per Gallon") |>
  save_tdr(my_mtcars, "my_mtcars_tabular-data-resource.yaml")
}
}
