% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{set_attrs}
\alias{set_attrs}
\alias{set_col_attrs}
\title{set attributes for an object}
\usage{
set_attrs(.x, ...)

set_col_attrs(.x, var, ...)
}
\arguments{
\item{.x}{an object to add attributes to}

\item{...}{additional arguments with name and value corresponding to metadata name and value}

\item{var}{name of variable in .x to add attributes to}
}
\value{
.x with the updated attributes
}
\description{
attributes are added onto existing attributes
}
\examples{
set_attrs(mtcars, name = "Motor Trend Cars", year = "1974")
set_col_attrs(mtcars, mpg, name = "MPG", description = "Miles Per Gallon")
}
