% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg_s3.R
\name{stow_codec_dpkg}
\alias{stow_codec_dpkg}
\title{Get a dpkg from the public CoDEC S3 bucket}
\usage{
stow_codec_dpkg(codec_dpkg)
}
\arguments{
\item{x}{name of CoDEC dpkg}

\item{dir}{directory in which to write the data package}
}
\value{
path to downloaded datapackage directory
}
\description{
Public data packages are downloaded from \verb{s3://geomarker-io/codec_data}
}
\examples{
dpkg_s3_get("drivetime") |>
  dpkg_read()
}
