% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pins.R
\name{codec_board_local_dev}
\alias{codec_board_local_dev}
\alias{write_codec_pin}
\title{Write a CoDEC data table to the local CoDEC board directory}
\usage{
codec_board_local_dev()

write_codec_pin(x)
}
\arguments{
\item{x}{a codec_tbl object created with as_codec_tbl()}
}
\description{
This function is used by developers creating and updating CoDEC data
tables.
}
\keyword{internal}
