% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_metadata}
\alias{write_metadata}
\title{save (table- and column-specific) metadata to a file}
\usage{
write_metadata(.x, file = "tabular-data-resource.yaml")
}
\description{
save (table- and column-specific) metadata to a file
}
