% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{write_tdr}
\alias{write_tdr}
\title{extract data resource metadata from a data frame and save it to a file}
\usage{
write_tdr(.x, file = "tabular-data-resource.yaml", codec = TRUE)
}
\arguments{
\item{.x}{a data.frame or tibble}

\item{file}{name of yaml file to write metadata to}

\item{codec}{logical; include only CODEC properties or schema? (see \code{?codec_tdr} for details)}
}
\value{
.x (invisibly)
}
\description{
extract data resource metadata from a data frame and save it to a file
}
\examples{
\dontrun{
mtcars |>
  add_attrs(name = "Motor Trend Cars", year = "1974") |>
  add_col_attrs(mpg, title = "MPG", description = "Miles Per Gallon") |>
  add_type_attrs() |>
  save_tdr(my_mtcars, "my_mtcars_tabular-data-resource.yaml")
}
}
