% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_tdr.R
\name{write_tdr_csv}
\alias{write_tdr_csv}
\title{write a tabular-data-resource yaml file and data csv file based on a data.frame or tibble}
\usage{
write_tdr_csv(.x, dir = getwd(), codec = TRUE)
}
\arguments{
\item{.x}{data.frame or tibble with metadata attributes (at least a 'name' attribute)}

\item{dir}{path to directory where tdr will be created; see details}

\item{codec}{logical; use only CoDEC properties?}
}
\value{
.x, invisibly
}
\description{
The \code{path} argument specifies where the folder containing
the codec-tdr will be created.  Within this path, the folder
for the codec-tdr will be named based on the name attribute
of the data.frame or tibble. The CSV data file will be named
based on the name attribute of the data.frame or tibble
and a "tabular-data-resource.yaml" file will also be created.
}
