% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{add_type_attrs}
\alias{add_type_attrs}
\title{automatically add "type" attributes for all columns in a data.frame based on their class}
\usage{
add_type_attrs(.x)
}
\description{
automatically add "type" attributes for all columns in a data.frame based on their class
}
