% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg.R
\name{as_codec_dpkg}
\alias{as_codec_dpkg}
\title{as_codec_dpkg}
\usage{
as_codec_dpkg(x, name, version, title = character(), homepage = character())
}
\arguments{
\item{x}{data.frame or tibble meeting CoDEC data specifications above}
}
\value{
a dpkg object
}
\description{
Convert a tibble to a data package (\code{dpkg}) object in R while checking it
against CoDEC data specifications:
}
\details{
\enumerate{
\item The data must include a \href{https://www2.census.gov/geo/pdfs/education/CensusTracts.pdf}{census tract}
identifier column (i.e., \verb{census_tract_id_\{2000,2010,2020\}}).
The column must contain 11-digit
\href{https://www.census.gov/programs-surveys/geography/guidance/geo-identifiers.html}{GEOID}
identifiers for every census tract in Hamilton County, OH.
\item The data includes a year column (\code{year}), an integer year representing the
vintage of the data (e.g. "2020").
The data can optionally include a month column (\code{month}), an integer month of the year.
}

Data must be structured in a tidy format such that each row is an observation
for a specific census tract at a specific year (and month).

Use this function to check the structure of a CoDEC data package before
writing it to disk or uploading it to a repository.
}
