% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_codec_tdr}
\alias{check_codec_tdr}
\title{check CODEC tabular-data-resource}
\usage{
check_codec_tdr(tdr)
}
\arguments{
\item{tdr}{a codec tabular-data-resource list object}
}
\description{
...
}
