% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_codec_tdr_csv}
\alias{check_codec_tdr_csv}
\title{check files, check metadata, check data}
\usage{
check_codec_tdr_csv(.x)
}
\arguments{
\item{.x}{path to folder containing the tdr}
}
\description{
check files, check metadata, check data
}
