% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_codec_tdr_csv}
\alias{check_codec_tdr_csv}
\alias{check_census_tract_id}
\alias{check_date}
\alias{check_files}
\alias{check_codec_tdr}
\alias{check_tdr_name}
\alias{check_tdr_path}
\title{Check a tabular-data-resource against CoDEC specifications}
\usage{
check_codec_tdr_csv(path)

check_census_tract_id(x)

check_date(x)

check_files(path)

check_codec_tdr(x)

check_tdr_name(name)

check_tdr_path(path)
}
\arguments{
\item{path}{path to tdr folder}

\item{x}{a codec fr_tdr object (or data frame for check_census_tract_id(), check_date()
and a list for check_codec_tdr())}

\item{name}{the name field from tabular-data-resource.yaml}
}
\value{
for \code{check_codec_tdr_csv}, a tibble with added
tabular-data-resource attributes (equivalent to read_tdr_csv with \code{codec = TRUE})
}
\description{
\code{check_codec_tdr_csv()} will use each of these functions to
check that
\itemize{
\item the file structure is correct
\item the metadata can be read using CoDEC specifications
\item the CSV data can be read in with accompanying metadata
\item the data contains a census tract identifier column
\item the data contains a year (or year and month) column(s)
\item all fields in the CSV data are described in the metadata and vice-versa
See \code{vignette("codec-specs")} for the CoDEC specifications.
}
}
