% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_files}
\alias{check_files}
\title{Check files}
\usage{
check_files(.x)
}
\arguments{
\item{.x}{path to folder containing the tdr}
}
\value{
.x, invisibly
}
\description{
Errors will be raised if the CODEC tabular data resource does not meet the following requirements:
\itemize{
\item tdr directory must contain a \code{.csv} with the same name and a \code{tabular-data-resource.yaml} file
\item both files must use UTF-8 character encoding and have newlines encoded as \verb{\\n} or \verb{\\r\\n}
\item the data file must:
\itemize{
\item follow the \href{https://www.rfc-editor.org/rfc/rfc4180}{RFC 4180 standard} for CSV files \emph{and}
\item have a header row containing a \emph{unique} name for each field
\item use \code{.} as the decimal mark (e.g., \code{1.34})
\item not use a grouping mark (e.g., \code{1200} instead of \verb{1,200})
}
}
}
