% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_data.R
\name{check_files}
\alias{check_files}
\title{Check for files}
\usage{
check_files(.x)
}
\arguments{
\item{.x}{path to folder containing the tdr}
}
\value{
.x, invisibly
}
\description{
Errors will be raised if the CODEC tabular data resource does not meet the following requirements:
}
\details{
\itemize{
\item tdr directory must exist and have a matching CSV data file and a tabular-data-resource.yaml file
\item both files must use UTF-8 character encoding and have newlines encoded as \verb{\\n} or \verb{\\r\\n}
\item the data file must end in \code{.csv}, have a header row containing the unique name of each field,
and be able to be read using the \href{https://www.rfc-editor.org/rfc/rfc4180}{RFC 4180 standard} for CSV files
}
}
