% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_check.R
\name{check_tdr_name}
\alias{check_tdr_name}
\title{Check for Tabular Data Resource Name}
\usage{
check_tdr_name(name)
}
\arguments{
\item{name}{the name field from tabular-data-resource.yaml}
}
\value{
an error message if conditions are not met
}
\description{
Name must be an identifier string composed of lower
case alphanumeric characters, _, -, and .
}
\examples{
check_tdr_name("name")
}
