% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_addr_geo}
\alias{cincy_addr_geo}
\title{Cincy address geographies}
\usage{
cincy_addr_geo()
}
\value{
a simple features object with columns \code{cagis_address}, \code{cagis_address_place}, \code{cagis_address_type},
\code{cagis_s2}, \code{cagis_parcel_id}, \code{cagis_is_condo}, and a geometry column (\code{s2_geography})
}
\description{
CAGIS data (see \code{install_cagis_data()}) provides a list of all addresses in Hamilton County.
Addresses are filtered for the following criteria:
\itemize{
\item use only addresses that have \code{STATUS} of \code{ASSIGNED}, \code{USING}, or \code{REGISTERED} and are not orphaned (\code{ORPHANFLG == "N"})
\item omit addresses with \code{ADDRTYPE}s that are milemarkers (\code{MM}), parks (\code{PAR}), infrastructure projects (\code{PRJ}),
cell towers (\code{CTW}), vacant or commercial lots (\code{LOT}), and other miscellaneous non-residential addresses (\code{MIS}, \code{RR}, \code{TBA})
\item s2 cell is derived from LONGITUDE and LATITUDE fields in CAGIS address database
}
}
\examples{
cincy_addr_geo()
}
