% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_census_geo}
\alias{cincy_census_geo}
\title{cincy_census_geo()
cincy_census_geo("bg", "2019", s2_geography = FALSE)}
\usage{
cincy_census_geo(
  geography = c("tract", "bg"),
  vintage = as.character(2024:2013)
)
}
\description{
cincy_census_geo()
cincy_census_geo("bg", "2019", s2_geography = FALSE)
}
