% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_zcta_geo}
\alias{cincy_zcta_geo}
\title{cincy_zcta_geo()
cincy_zcta_geo("2018")}
\usage{
cincy_zcta_geo(vintage = as.character(2024:2013))
}
\description{
cincy_zcta_geo()
cincy_zcta_geo("2018")
}
