% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{cincy_zcta_geo}
\alias{cincy_zcta_geo}
\title{Cincy ZIP Code Tabulation Areas}
\usage{
cincy_zcta_geo(vintage = as.character(2024:2013))
}
\arguments{
\item{vintage}{a character vector of a year corresponding to the vintage of TIGER/Line data}
}
\value{
a simple features object with a geographic identifier column (\code{geoid})
and a geometry column (\code{s2_geography})
}
\description{
Read \href{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/zctas.html}{ZIP Code Tabulation Areas (ZCTAs)}
geographies from the online Census
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{TIGER/Line}
files into R
}
\examples{
cincy_zcta_geo()
cincy_zcta_geo("2018")
}
