% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDEC-package.R
\docType{package}
\name{codec-package}
\alias{codec}
\alias{codec-package}
\title{About the CoDEC package}
\description{
The CoDEC data package specifications provide a set of patterns designed to make sharing tabular community-level data easier.
Examples of this type of data include the pediatric hospitalization rate per month per census tract, the total number of gunshots per season per neighborhood, and the housing code enforcement density per year per ZIP code.

Data is specified as average values or total counts for census tract geographies during a specific year (or year and month).
The required \strong{census tract identifer} and \strong{year} (or \emph{year} and \emph{month})
columns in a CoDEC table contain the spatiotemporal information that can be used to link other data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geomarker.io/codec}
  \item Report bugs at \url{https://github.com/geomarker-io/codec/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com}

Authors:
\itemize{
  \item Erika Manning
  \item Andrew Vancil
}

}
\keyword{internal}
