% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDEC-package.R
\docType{package}
\name{CoDEC-package}
\alias{CoDEC}
\alias{CoDEC-package}
\title{CoDEC: Community Data Explorer for Cincinnati}
\description{
CoDEC provides tools for working with metadata in R and storing it alongside data in a YAML file. This package serves as the definition of the CoDEC data specifications and provides helpers to contribute and validate CoDEC data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geomarker-io/CoDEC}
  \item \url{http://geomarker.io/CoDEC/}
  \item Report bugs at \url{https://github.com/geomarker-io/CoDEC/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com}

Authors:
\itemize{
  \item Erika Manning
}

}
\keyword{internal}
