% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-package.R
\docType{package}
\name{codec-package}
\alias{codec}
\alias{codec-package}
\title{codec: Community Data Explorer for Cincinnati}
\description{
codec provides tools for working with metadata in R and storing it alongside data in a YAML file. This package serves as the definition of the CoDEC data specifications and provides helpers to contribute and validate CoDEC data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geomarker-io/codec}
  \item \url{http://geomarker.io/codec/}
  \item Report bugs at \url{https://github.com/geomarker-io/codec/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com}

Authors:
\itemize{
  \item Erika Manning
  \item Andrew Vancil
}

}
\keyword{internal}
