% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_interpolate.R
\name{codec_as_sf}
\alias{codec_as_sf}
\title{Coerce CoDEC data package into a simple features object}
\usage{
codec_as_sf(x)
}
\arguments{
\item{x}{a CoDEC data package}
}
\value{
a simple features object with a geographic identifier column (\code{geoid})
and a geometry column (\code{s2_geography}) in addition to the existing columns in x
}
\description{
This functions uses the name of the census tract column in
the CoDEC data package to add the appropriate \code{cincy_census_geo()}
and convert it into a sf object.
}
\details{
Tract identifers do not change across decennial censuses, but the digital representation of their boundaries
may be improved over time.  Here, data packages using 2010 tract identifers use the TIGER/Line 2019 tract shapefiles
and data packages using 2020 tract identifiers use the TIGER/Line 2020 tract shapefiles
}
\examples{
codec_as_sf(get_codec_dpkg("property_code_enforcements-v0.2.0"))
}
