% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_design.R
\name{codec_colors}
\alias{codec_colors}
\title{CoDEC colors}
\usage{
codec_colors(n = NULL)
}
\arguments{
\item{n}{a numeric vector of color numbers or character vector of color names; if NULL returns named vector of available colors}
}
\description{
CoDEC colors
}
