% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_tdr.R
\name{codec_data}
\alias{codec_data}
\title{read CoDEC tabular data resource}
\usage{
codec_data(name, interpolate_to = NULL)
}
\arguments{
\item{name}{name of installed codec tabular data resource}

\item{interpolate_to}{a {cincy} geography object; codec data
will be returned at this geography}
}
\value{
a tibble (codec tabular data resource)
}
\description{
This function is shorthand for reading a CoDEC tabular
data resource using \code{read_tdr_csv()} from the installed
R package.
}
\details{
the {cincy} geography object is supplied to \code{interpolate_to}
is passed to \code{cincy::interpolate}, interpolating the tract-level codec
data to the supplied geography using population weights
}
\examples{
codec_data("hamilton_traffic")
}
