% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_tdr.R
\name{codec_data}
\alias{codec_data}
\title{read CoDEC tabular data resource}
\usage{
codec_data(name)
}
\arguments{
\item{name}{name of installed codec tabular data resource}
}
\value{
a tibble (codec tabular data resource), or simple features object when \code{geometry = TRUE}
}
\description{
This function is shorthand for reading a CoDEC tabular
data resource using \code{read_tdr_csv()} from the installed
R package.
}
\examples{
codec_data("hamilton_drivetime")
}
