% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg_s3.R
\name{codec_dpkg_s3_put}
\alias{codec_dpkg_s3_put}
\title{Put a dpkg into the public CoDEC S3 bucket}
\usage{
codec_dpkg_s3_put(x)
}
\arguments{
\item{x}{a data package (\code{dpkg::dpkg}) object}
}
\value{
character string URI of uploaded resource
}
\description{
The \href{https://aws.amazon.com/cli/}{AWS CLI} tool must be installed and authenticated to
write to \verb{s3://geomarker-io/codec_data}.
The resulting data package will be available publicly.
}
\examples{
\dontrun{
# use aws command line to login interactively via profile sso account"
system2("aws", c("sso", "login", "--profile", "geomarker-io"))
# make sure to set AWS_PROFILE so the AWS CLI tool knows to use it by default"
Sys.setenv("AWS_PROFILE" = "geomarker-io")
}
}
