% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_interpolate.R
\name{codec_interpolate}
\alias{codec_interpolate}
\title{Spatially interpolate community-level data}
\usage{
codec_interpolate(from, to = NULL, weights = c("pop", "homes", "area"))
}
\arguments{
\item{from}{a CoDEC data package}

\item{to}{; if NULL, tract geographies are returned without data interpolation}
}
\value{
a simple features object with a geographic identifier column (\code{geoid})
and a geometry column (\code{s2_geography}) in addition to the (interpolated) columns in \code{from}
}
\description{
Weights at the census block-level are used to spatially interpolate different geographies.
Block-level total population, total number of homes, or total land area from the 2020 Census
can be chosen to use for the weights.
}
\details{
Tract identifers do not change across decennial censuses, but the digital representation of their boundaries
may be improved over time.  Here, data packages using 2010 tract identifers use the TIGER/Line 2019 tract shapefiles
and data packages using 2020 tract identifiers use the TIGER/Line 2020 tract shapefiles
}
\examples{
codec_interpolate(get_codec_dpkg("property_code_enforcements-v0.2.0"))
codec_interpolate(get_codec_dpkg("property_code_enforcements-v0.2.0"), to = "zcta")
}
