% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codec_latest_annual}
\alias{codec_latest_annual}
\title{CoDEC latest annual}
\format{
An object of class \code{codec_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 226 rows and 66 columns.
}
\usage{
codec_latest_annual
}
\description{
CoDEC data is harmonized as the most recently available annual (or annual average of monthly) values
at the census tract 2020 geography.
The year represents the year this table was assembled and mainly used as a placeholder to this object
can be used with functions that take a codec_tbl object (see \code{?as_codec_tbl}).
See the description metadata for the actual latest years used for each CoDEC table.
}
\examples{
tibble::glimpse(codec_latest_annual)
glue::glue(attr(codec_latest_annual, "description"))
}
\keyword{datasets}
