% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pins.R
\name{codec_read}
\alias{codec_read}
\title{Read a CoDEC table from the online data catalog}
\usage{
codec_read(name, board = codec_board())
}
\arguments{
\item{name}{name of CoDEC table in the online CoDEC data catalog}

\item{board}{codec board to read from}
}
\value{
a codec_tbl object (see \code{as_codec_tbl()})
}
\description{
This function uses the pins package to read from the online CoDEC data catalog,
ensuring that metadata is present in the returned codec_tbl object.
}
\details{
Read from the online data without installing this package directly
using the pins package with \code{codec_board()}.
}
\examples{
d <- codec_read("traffic")
head(d)
attr(d, "title")
message(attr(d, "description"))
}
