% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pins.R
\name{codec_read}
\alias{codec_read}
\alias{codec_list}
\alias{codec_board}
\title{CoDEC online data catalog}
\usage{
codec_read(name, board = codec_board())

codec_list(board = codec_board())

codec_board(
  version = paste0("v", utils::packageVersion("codec")),
  cache = NULL,
  use_cache_on_failure = rlang::is_interactive(),
  headers = NULL
)
}
\arguments{
\item{name}{The name of the CoDEC table in the online CoDEC data catalog.}

\item{board}{a pins board object; create with \code{codec_board()} to read from earlier versions of the catalog
or to change the caching behavior of the pins package}

\item{version}{specify a version of the online data catalog using a commit SHA, tag, or branch of geomarker-io/codec;
syncs with the version of the installed package by default}

\item{cache}{Cache path. Every board requires a local cache to avoid
downloading files multiple times. The default stores in a standard
cache location for your operating system, but you can override if needed.}

\item{use_cache_on_failure}{If the pin fails to download, is it ok to
use the last cached version? Defaults to \code{is_interactive()} so you'll
be robust to poor internet connectivity when exploring interactively,
but you'll get clear errors when the code is deployed.}

\item{headers}{Named character vector for additional HTTP headers (such as for
authentication). See \code{\link[pins:connect_auth_headers]{connect_auth_headers()}} for Posit Connect support.}
}
\value{
For \code{codec_read()}, a codec_tbl object (see \code{as_codec_tbl()})

For \code{codec_list()}, a character vector CoDEC table names

For \code{codec_board()}, a pins_board object
}
\description{
The CoDEC online data catalog is hosted on GitHub alongside
the source code for this package.
\itemize{
\item Use \code{codec_read()} as a shortcut to read a CoDEC table
into R as a codec_tbl object (see \code{?as_codec_tbl})
\item Use \code{codec_list()} as a shortcut to list available CoDEC table pins
\item \code{codec_board()} can be used to create a pin board
object (see \code{?pins::pins}) based on a specific version of the codec package
}
}
\details{
The pin for each CoDEC table has versions (see \code{?pins::pin_versions}),
but \code{codec_board()} can be used to specify a state of the online data catalog
based on the version of the codec package. (See examples)
}
\examples{
# list available CoDEC tables
codec_list()

# read a CoDEC table and inspect its metadata
d <- codec_read("traffic")
head(d)
attr(d, "title")
message(attr(d, "description"))

# change the defaults for codec_board() to read from older versions of the board
codec_board() |>
  pins::pin_versions("crime")
codec_board("v3.0.0-rc1") |>
  pins::pin_versions("crime")
}
