% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pin.R
\name{codec_read}
\alias{codec_read}
\title{use pins to read from the CoDEC pins board, transforming the object back into a CoDEC tbl}
\usage{
codec_read(name)
}
\arguments{
\item{name}{name of CoDEC table in the online CoDEC data catalog}
}
\value{
a codec_tbl object (see \code{as_codec_tbl()})
}
\description{
This function uses the pins package to read from the online CoDEC data catalog,
ensuring that metadata is present in the returned codec_tbl object.
Read from the online data without installing this package directly
using {pins} with \code{codec_board()}.
}
\examples{
d <- codec_read("traffic")
head(d)
attr(d, "title")
attr(d, "description")
}
