% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_tdr_tables.R
\name{codec_tdr}
\alias{codec_tdr}
\title{return a list representing a valid CODEC-specific tabular-data-resource structure}
\usage{
codec_tdr()
}
\description{
return a list representing a valid CODEC-specific tabular-data-resource structure
}
\examples{
codec_tdr()
}
