% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_tdr.R
\name{codec_tdr}
\alias{codec_tdr}
\title{codec_tdr}
\usage{
codec_tdr()
}
\value{
a list of named character vectors, one for each of "property", "schema", and "fields"
}
\description{
Defines the list of names (and descriptions) of valid CODEC \emph{property}, \emph{schema}, and \emph{fields} descriptors
}
\examples{
codec_tdr()$property
codec_tdr()$schema
codec_tdr()$fields
}
