% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CODECtools-package.R
\docType{package}
\name{CODECtools-package}
\alias{CODECtools}
\alias{CODECtools-package}
\title{CODECtools: Support CODEC data infrastructure}
\description{
CODECtools provides tools for working with metadata in R and storing it alongside data in a YAML file. This package serves as the definition of the CODEC data specifications and provides helpers to contribute and validate CODEC data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geomarker-io/CODECtools}
  \item \url{http://geomarker.io/CODECtools/}
  \item Report bugs at \url{https://github.com/geomarker-io/CODECtools/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com}

Authors:
\itemize{
  \item Erika Manning
}

}
\keyword{internal}
