% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg.R
\name{dpkg_read}
\alias{dpkg_read}
\title{Read a data package from a folder}
\usage{
dpkg_read(dpkg, return = c("rds", "path", "metadata", "md"))
}
\arguments{
\item{dpkg}{path or url to folder containing a datapackage.yaml file}

\item{return}{what to return? "rds" returns the R object by reading in the rds file;
"path" returns the path to the rds file; "metadata" (or "md") returns a list of the metadata
stored with the rds file in the data package with \code{dpkg_write()}}
}
\value{
either the R object, path to the rds file, or a list of metadata
}
\description{
Read a \href{https://datapackage.org/standard/data-package/}{data package} with
one \href{https://datapackage.org/standard/data-resource/}{data resource} that
is a RDS object created with \code{dpkg_write()}.
}
