% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg.R
\name{dpkg_s3_get}
\alias{dpkg_s3_get}
\title{Get a dpkg file from the CoDEC S3 bucket under \verb{/data}}
\usage{
dpkg_s3_get(x = "aadt", dir)
}
\arguments{
\item{x}{name of CoDEC dpkg}

\item{dir}{directory in which to write the data package}
}
\description{
Get a dpkg file from the CoDEC S3 bucket under \verb{/data}
}
