% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_s3.R
\name{dpkg_s3_get}
\alias{dpkg_s3_get}
\title{Get a dpkg from the public CoDEC S3 bucket}
\usage{
dpkg_s3_get(x, dir = tools::R_user_dir("codec", "data"))
}
\arguments{
\item{x}{name of CoDEC dpkg}

\item{dir}{directory in which to write the data package}
}
\value{
path to downloaded datapackage directory
}
\description{
Public data packages are downloaded from \verb{s3://geomarker-io/codec_data}
}
\examples{
dpkg_s3_get("drivetime") |>
  dpkg_read()
}
