% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpkg_s3.R
\name{dpkg_s3_put}
\alias{dpkg_s3_put}
\title{Put a dpkg into the CoDEC S3 bucket under \verb{/data}}
\usage{
dpkg_s3_put(x)
}
\arguments{
\item{x}{path to datapackage.yaml file}
}
\description{
Ensure {paws} can connect via the usual methods.
See the developer guide at https://www.paws-r-sdk.com/developer_guide/credentials/.
}
\examples{
\dontrun{
# use credentials in .env file to authenticate with aws
library(dotenv)
# alternatively, use aws command line to login interactively via profile sso account
system2("aws", c("sso", "login", "--profile", "geomarker-io"))
# make sure to set AWS_PROFILE so {paws} can authenticate
Sys.setenv("AWS_PROFILE" = "geomarker-io")
}
}
