% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg_s3.R
\name{get_codec_dpkg}
\alias{get_codec_dpkg}
\alias{stow_codec_dpkg}
\title{Read a dpkg from the public CoDEC repository into R}
\usage{
get_codec_dpkg(codec_dpkg, overwrite = FALSE)

stow_codec_dpkg(codec_dpkg, overwrite = FALSE)
}
\arguments{
\item{codec_dpkg}{name of CoDEC dpkg}

\item{overwrite}{logical; re-download the remote file even though
a local file with the same name exists?}
}
\value{
for \code{get_codec_dkg()}, an data package (\code{dpkg::as_dpkg()}) object; for
\code{stow_codec_dpkg()}, the path to the downloaded data package parquet file
}
\description{
Public data packages are downloaded from \verb{s3://geomarker-io/codec_data}.
\code{dpkg::stow()} is used to cache a local copy in the user's data directory so that
it is available later without having to redownload it.
}
\examples{
get_codec_dpkg("traffic-v0.1.0")

stow_codec_dpkg("traffic-v0.1.0")
}
