% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincy_geographies.R
\name{install_cagis_data}
\alias{install_cagis_data}
\title{Install CAGIS GIS database}
\usage{
install_cagis_data(
  cagis_data_url =
    "https://www.cagis.org/Opendata/Quarterly_GIS_Data/CAGISOpenDataQ1_2025.gdb.zip"
)
}
\arguments{
\item{cagis_data_url}{the url to the CAGIS Open Data .gdb.zip file; this changes quarterly, so
\href{https://www.cagis.org/Opendata/Quarterly_GIS_Data}{check} for something more recent if the file cannot be found}
}
\description{
This installs the CAGIS Open Data GIS database (\code{.gdb}) into the data
directory for the codec package. Once downloaded, it will be reused
across R sessions on the same computer.
The geodatabase contains many
\href{https://www.cagis.org/Opendata/Quarterly_GIS_Data/OpenData_Layer_List.txt}{layers} that are
updated quarterly. (Historical geodatabases are not available here.)
}
\examples{
\dontrun{
install_cagis_data()
sf::st_layers(install_cagis_data())$name
}
}
\seealso{
This function is called by \code{cincy_neighborhood_geo()}, \code{cincy_city_geo()}
and others that import individual layers.
}
