% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-resource.R
\name{make_attr_from_tdr}
\alias{make_attr_from_tdr}
\title{add attributes in a metadata list to a data.frame}
\usage{
make_attr_from_tdr(.x, metadata)
}
\arguments{
\item{.x}{a data.frame or tibble}
}
\value{
frictionless metadata as a list
}
\description{
add attributes in a metadata list to a data.frame
}
