% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{make_tdr_from_attr}
\alias{make_tdr_from_attr}
\title{make metadata list from attributes of data.frame}
\usage{
make_tdr_from_attr(.x)
}
\arguments{
\item{.x}{a data.frame or tibble}
}
\value{
frictionless metadata as a list
}
\description{
make metadata list from attributes of data.frame
}
