% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{read_codec}
\alias{read_codec}
\title{read a CODEC tabular-data-resource from the cloud}
\usage{
read_codec(name, force = FALSE)
}
\arguments{
\item{name}{name of CODEC tabular data resource}

\item{force}{ignore existing data and redownload data?}
}
\value{
tibble with added tabular-data-resource attributes
}
\description{
If not available locally, the CODEC tabular-data-resource will be
downloaded from \verb{s3://codec-data/}.
}
