% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular-data-resource.R
\name{read_tdr}
\alias{read_tdr}
\title{read metadata in from a tabular-data-resource.yaml file}
\usage{
read_tdr(file = "tabular-data-resource.yaml")
}
\arguments{
\item{file}{name of yaml file to write metadata to}
}
\value{
a list of frictionless metadata
}
\description{
read metadata in from a tabular-data-resource.yaml file
}
