% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_tdr.R
\name{read_tdr}
\alias{read_tdr}
\title{Read a tabular data resource into R from disk or the web}
\usage{
read_tdr(tdr_location)
}
\arguments{
\item{tdr_file}{path or url to a \code{tabular-data-resource.yaml} file
(or a directory containing a \code{tabular-data-resource.yaml} file)}
}
\value{
a list of (1) the "tdr" and (2) the path to its "csv_file"
}
\description{
In addition, the path (or url) of the CSV is returned.
}
\examples{
read_tdr("https://github.com/geomarker-io/hamilton_landcover/releases/download/v0.1.0") |>
str(4)
}
