% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{read_tdr_csv}
\alias{read_tdr_csv}
\title{read a CSV tabular data resource into R}
\usage{
read_tdr_csv(file = "tabular-data-resource.yaml")
}
\arguments{
\item{file}{path to tabular-data-resource yaml file}
}
\value{
tibble with added tabular-data-resource attributes
}
\description{
The CSV file defined in a tabular-data-resource yaml file
will be read into R using \code{readr::read_csv}. Columns to
read and their types are set using metadata. Metadata
(descriptors and schema) are stored as attributes
of the returned tibble.
}
\details{
\emph{Note:}
\itemize{
\item the \code{path} descriptor is always relative to the tabular-data-resource file
\item descriptors will always be restricted to those in \code{codec_tdr()}
}
}
