% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{read_tdr_csv}
\alias{read_tdr_csv}
\title{read a CSV tabular data resource into R}
\usage{
read_tdr_csv(dir = getwd(), codec = TRUE, ...)
}
\arguments{
\item{dir}{path or connection to folder that contains a
tabular-data-resource.yaml file}

\item{codec}{logical; use only CODEC descriptors?}

\item{...}{additional options passed onto \code{readr::read_csv()}}
}
\value{
tibble with added tabular-data-resource attributes
}
\description{
The CSV file defined in a tabular-data-resource yaml file
are read into R using \code{readr::read_csv()}. Metadata
(descriptors and schema) are stored as attributes
of the returned tibble and are also used to set
the column classes of the returned data.frame or tibble.
}
