% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_tdr.R
\name{read_tdr_csv}
\alias{read_tdr_csv}
\title{read a CSV tabular data resource into R from disk or the web}
\usage{
read_tdr_csv(tdr_file, codec = TRUE, ...)
}
\arguments{
\item{tdr_file}{path or url to a \code{tabular-data-resource.yaml} file
(or a directory containing a \code{tabular-data-resource.yaml} file)}

\item{codec}{logical; use only CoDEC properties?}

\item{...}{additional options passed onto \code{readr::read_csv()}}
}
\value{
tibble with added tabular-data-resource attributes
}
\description{
The CSV file defined in a tabular-data-resource yaml file
is read into R using \code{readr::read_csv()}. Metadata
(properties and schema) are stored as attributes
of the returned tibble and are also used to set
the column classes of the returned data.frame or tibble.
Files starting with \verb{http://} or \verb{https://} will be automatically
downloaded locally first.
}
