% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdr.R
\name{release_codec_tdr}
\alias{release_codec_tdr}
\title{experimental!}
\usage{
release_codec_tdr(.x, version = "0.1.0", gh = TRUE)
}
\arguments{
\item{.x}{a CODEC tdr}

\item{version}{the version number to be used for the release}

\item{gh}{logical; also create and upload CODEC tdr to a new GitHub release?}
}
\description{
experimental!
}
