% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_dpkg_s3.R
\name{stow_codec_dpkg}
\alias{stow_codec_dpkg}
\title{Get a dpkg from the public CoDEC S3 bucket}
\usage{
stow_codec_dpkg(codec_dpkg)
}
\arguments{
\item{codec_dpkg}{name of CoDEC dpkg}
}
\value{
path to downloaded data package parquet file
}
\description{
Public data packages are downloaded from \verb{s3://geomarker-io/codec_data}
}
\examples{
\dontrun{
  dpkg_s3_get("drivetime-v0.1.0") |>
    dpkg::read_dpkg()
}
}
