% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec_pin.R
\name{write_codec_pin}
\alias{write_codec_pin}
\title{Write a CoDEC data table to the CoDEC board directory}
\usage{
write_codec_pin(x)
}
\arguments{
\item{x}{a codec_tbl object created with as_codec_tbl()}
}
\description{
This function is used by developers creating and updating CoDEC data tables.
}
